Split implementation of FTP API out from its interface.
